fix: fixed environment temperature with scale

+++ src/main/java/com/ryan/project/smarthomehub/module/device/Heater.java
@@ -91,7 +91,7 @@ public class Heater extends Device implements TemperatureSetting, OnOff, DeviceS
             BigDecimal environmentTemperature = jsonElement.getAsJsonObject().get("environment_temperature").getAsBigDecimal();
             String oldTemperature = documentReference.get().get().get("states.thermostatTemperatureAmbient", String.class);
             if (new BigDecimal(oldTemperature).intValue()!=(environmentTemperature.intValue())) {
-                documentReference.update("states.thermostatTemperatureAmbient", environmentTemperature);
+                documentReference.update("states.thermostatTemperatureAmbient", environmentTemperature.intValue());
             }
         } catch (Exception e) {
             log.error("error update environment temperature", e);
